requirements.txt
auth0-python==3.16.0
python-jose==3.3.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
blis==0.7.9
breadability==0.1.20
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
cryptography==40.0.2
cymem==2.0.7
docopt==0.6.2
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.5.0/es_core_news_sm-3.5.0-py3-none-any.whl
Flask==2.3.2
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
langcodes==3.3.0
lxml==4.9.2
MarkupSafe==2.1.2
multidict==6.0.4
murmurhash==1.0.9
networkx==2.6.3
nltk==3.8.1
numpy==1.24.3
openai==0.27.6
packaging==23.1
pandas==1.4.5
pathy==0.10.1
pdf2image==1.16.3
pdfminer.six==20211012
pdfplumber==0.9.0
Pillow==9.5.0
preshed==3.0.8
pycountry==22.3.5
pycparser==2.21
pydantic==1.10.7
PyPDF2==1.26.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.5.0
pytz==2023.3
regex==2023.5.5
requests==2.30.0
six==1.16.0
smart-open==6.3.0
soupsieve==2.4.1
spacy==3.5.2
srsly==2.4.6
sumy==0.11.0
